+++ bench/RowwiseAdagradFusedBenchmark.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) Facebook, Inc. and its affiliates.
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
  * All rights reserved.
  * This source code is licensed under the BSD-style license found in the
  * LICENSE file in the root directory of this source tree.
@@ -58,13 +58,13 @@ void run_benchmark(
   // normal_distribution<float> h_w_distribution;
 
   // TODO: check appropriate vals for g,h,w
-  for (int i = 0; i < g.size(); ++i) {
+  for (size_t i = 0; i < g.size(); ++i) {
     g[i] = 4 + i; // h_w_distribution(generator);
   }
-  for (int i = 0; i < h.size(); ++i) {
+  for (size_t i = 0; i < h.size(); ++i) {
     h_ref[i] = h[i] = 2 + i; // h_w_distribution(generator);
   }
-  for (int i = 0; i < w.size(); ++i) {
+  for (size_t i = 0; i < w.size(); ++i) {
     w_ref[i] = w[i] = 3 + i; // h_w_distribution(generator);
   }
